pyproject.toml
[project]
name = "factorio-icon-generator"
version = "0.1.0"
description = "Generates standalone icons"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["click>=8.1.7", "pillow>=11.0.0"]
